se realizan cambios en Migrations y Modelos
Los cambios se realizan en los tipos de migrations

Create-numero-fam-atendidas
create-detalle-nino
create-nino-desnutricion

Modelos
detallenino
ninodesnutricion
numerofamatendidas

+++ migrations/20240618172019-create-nino-desnutricion.js
@@ -10,10 +10,10 @@ module.exports = {
         type: Sequelize.INTEGER
       },
       noDesnutridos: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DOUBLE
       },
       desnutridos: {
-        type: Sequelize.INTEGER
+        type: Sequelize.DOUBLE
       },
       createdAt: {
         allowNull: false,
